+++ src/cone/app/tests/test_app.py
@@ -111,7 +111,8 @@ class TestApp(NodeTestCase):
             'cone.admin_password': 'admin',
             'cone.auth_secret': '12345',
             'cone.auth_reissue_time': '300',
-            'cone.auth_max_age': '600'
+            'cone.auth_max_age': '600',
+            'cone.plugins': 'cone.app.tests'  # ensure dummy main hooks called
         }
 
         # main